Set password correctly for remote user
There were a few problems. First, some leading whitespace was
sneaking into the password when it was passed as a parameter.
Second, the 'echo' was not passing the password correctly to
chpasswd, so the password never got set.

The whitespace is now being stripped. I use a 'here' command
to set the password instead of 'echo'. The password can now
also be changed remotely for an existing account.

Signed-off-by: Joel Nider <joel@ece.ubc.ca>