Copy key to /tmp for local account
When adding a new user to the local machine, the script fails to
install the ssh key because it is not in the expected location.
The script expects the key to be in /tmp because that is where it
is copied while setting up an account on a remote machine.

Copy the key to /tmp so the common parts of the script can find it.

Signed-off-by: Joel Nider <joel@ece.ubc.ca>