Don't add the same key twice to authorized_keys
When modifying an account (such as adding 'sudo' access) it is
possible that the key is passed as a parameter but has already been
added to the account. In that case, it will be added multiple times
to the authorized_keys file, increasing its size and making it
harder to maintain, for no reason.

Check to see if the key has already been added before adding it
again. The check is done by key name (the 3rd field, separated by
spaces), so it is possible to trick this check in a couple of ways.
First, if you use the same key but with a different name, the check
will not recognize it as the same key. This can be avoided by
checking the key type and value, but that seems like too much work
to avoid casual forgetfulness/carelessness. It will also only check
the first key in the file; it is possible to have multiple keys in
a key file; so any subsequent keys will be added, bypassing the check.
On the other hand, if the first key is a duplicate, none of the keys
will be added to authorized_keys. The lesson being, only use files
with a single key to avoid these problems.

Signed-off-by: Joel Nider <joel@ece.ubc.ca>