Check for the existence of 'authorized_keys'
Before adding the public key to the 'authorized_keys' file, we first try to
determine if the key has already been added to prevent duplicates. We do so
by grepping for the key. Grep will fail if the file doesn't exist, so we must
first make sure the file is there.

Signed-off-by: Joel Nider <joel@ece.ubc.ca>