Check for existence of user account
Before creating the account, we must first check to see if it already exists.
Otherwise, the 'useradd' function will fail. Before this patch, we relied on
the error code from 'useradd' to know if the account was already created but
this is unreliable. It seems different versions of 'useradd' return different
error codes (9 vs 130) and actually the script stops running for some reason,
probably because of commit 499eaa2ae82 which adds 'set -e' to exit on error.

To enhance reliability, we first call 'id -u' before 'useradd'.

Signed-off-by: Joel Nider <joel@ece.ubc.ca>