Fix a few bugs when calling remote script
Pass the password using single quotes so as to accept passwords that
contain $, which otherwise would represent a bash variable name.

Strip the path from the key file name. We already copied the file to
the /tmp directory so it is incorrect to use the original file path.

Pass 'verbose' through to the remote script for debugging purposes.

Signed-off-by: Joel Nider <joel.nider@unifabrix.com>